+++ src/olmo_core/data/__init__.py
@@ -8,11 +8,18 @@ Prepare your data by writing token IDs to numpy arrays on disk, using the
 `Dolma toolkit <https://allenai.github.io/dolma/>`_ for example.
 
 Configure and build your dataset using the :class:`~olmo_core.data.numpy_dataset.NumpyDatasetConfig`
-builder and pass it to :meth:`TrainerConfig.build() <olmo_core.train.TrainerConfig.build>`.
+builder, build your data loader with the :class:`~olmo_core.data.data_loader.NumpyDataLoaderConfig`
+builder, then pass it to :meth:`TrainerConfig.build() <olmo_core.train.TrainerConfig.build>`.
 """
 
 from .collator import DataCollator, PaddingDirection
-from .data_loader import DataLoaderBase, FSLDataLoader, VSLDataLoader
+from .data_loader import (
+    DataLoaderBase,
+    NumpyDataLoaderBase,
+    NumpyDataLoaderConfig,
+    NumpyFSLDataLoader,
+    NumpyVSLDataLoader,
+)
 from .mixes import DataMix
 from .numpy_dataset import (
     NumpyDatasetBase,
@@ -53,6 +60,8 @@ __all__ = [
     "DataCollator",
     "PaddingDirection",
     "DataLoaderBase",
-    "FSLDataLoader",
-    "VSLDataLoader",
+    "NumpyDataLoaderBase",
+    "NumpyFSLDataLoader",
+    "NumpyVSLDataLoader",
+    "NumpyDataLoaderConfig",
 ]