+++ src/olmo_core/train/callbacks/sequence_length_scheduler.py
@@ -3,7 +3,7 @@ import math
 from dataclasses import dataclass
 from typing import Any, Dict, Optional
 
-from olmo_core.data import NumpyFSLDataset
+from olmo_core.data import NumpyFSLDataLoader, NumpyFSLDataset
 from olmo_core.data.utils import melt_batch, truncate_batch
 from olmo_core.exceptions import OLMoConfigurationError
 from olmo_core.utils import gc_cuda
@@ -32,7 +32,8 @@ class SequenceLengthSchedulerCallback(Callback):
     during the warm-up is always a multiple of :data:`keep_multiple_of`.
 
     .. important::
-        This callback is only compatible with a :class:`~olmo_core.data.NumpyFSLDataset`.
+        This callback is only compatible with a :class:`~olmo_core.data.data_loader.NumpyFSLDataLoader`
+        training :data:`~olmo_core.train.Trainer.data_loader`.
 
     .. note::
         The "total tokens" recorded by the trainer and :class:`SpeedMonitorCallback` will
@@ -52,24 +53,24 @@ class SequenceLengthSchedulerCallback(Callback):
         if not self.enabled:
             return
 
-        if not isinstance(self.trainer.dataset, NumpyFSLDataset):
+        if not isinstance(self.trainer.data_loader, NumpyFSLDataLoader):
             raise OLMoConfigurationError(
-                "The sequence length scheduler callback requires a 'NumpyFSLDataset', "
-                f"got '{type(self.trainer.dataset)}' instead"
+                "The sequence length scheduler callback requires a 'NumpyFSLDataLoader', "
+                f"got '{type(self.trainer.data_loader)}' instead"
             )
 
+        dataset = self.trainer.data_loader.dataset
+        assert isinstance(dataset, NumpyFSLDataset)
+
         if self.truncate and (
-            self.trainer.dataset.sequence_length % self.min_sequence_length != 0
-            or (
-                math.log(self.trainer.dataset.sequence_length // self.min_sequence_length, 2) % 1
-                != 0
-            )
+            dataset.sequence_length % self.min_sequence_length != 0
+            or (math.log(dataset.sequence_length // self.min_sequence_length, 2) % 1 != 0)
         ):
             raise OLMoConfigurationError(
                 "train sequence length must be a multiple of 'min_sequence_length' by a power of 2 "
                 "when 'truncate=False'."
             )
-        elif self.trainer.dataset.sequence_length <= self.min_sequence_length:
+        elif dataset.sequence_length <= self.min_sequence_length:
             raise OLMoConfigurationError(
                 "train sequence length must be greater than 'min_sequence_length'"
             )
@@ -83,13 +84,15 @@ class SequenceLengthSchedulerCallback(Callback):
         if self.step > self.warmup_steps:
             return
 
-        assert isinstance(self.trainer.dataset, NumpyFSLDataset)
+        assert isinstance(self.trainer.data_loader, NumpyFSLDataLoader)
+        dataset = self.trainer.data_loader.dataset
+        assert isinstance(dataset, NumpyFSLDataset)
 
         new_seq_len: int
         if self.truncate:
             new_seq_len = _get_truncated_sequence_length(
                 self.min_sequence_length,
-                self.trainer.dataset.sequence_length,
+                dataset.sequence_length,
                 self.step,
                 self.warmup_steps,
                 self.keep_multiple_of,
@@ -103,7 +106,7 @@ class SequenceLengthSchedulerCallback(Callback):
         else:
             new_seq_len = _get_split_sequence_length(
                 self.min_sequence_length,
-                self.trainer.dataset.sequence_length,
+                dataset.sequence_length,
                 self.step,
                 self.warmup_steps,
             )
@@ -118,7 +121,7 @@ class SequenceLengthSchedulerCallback(Callback):
             # in each micro-batch.
             assert self._og_rank_microbatch_size is not None
             new_rank_microbatch_size = self._og_rank_microbatch_size * (
-                self.trainer.dataset.sequence_length // new_seq_len
+                dataset.sequence_length // new_seq_len
             )
             self.trainer.rank_microbatch_size = new_rank_microbatch_size