+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated default layer norm epsilon for OLMo models from `1e-5` to `1e-6` to match latest model.
+- Renamed `FSLDataLoader` to `NumpyFSLDataLoader`.
+- Renamed `VSLDataLoader` to `NumpyVSLDataLoader`.
+- The trainer now takes a `data_loader: DataLoaderBase` instead of a `dataset: NumpyDatasetBase`.
 
 ## [v1.3.2](https://github.com/allenai/OLMo-core/releases/tag/v1.3.2) - 2024-09-27